Move ProtectedRoute into /components and update outdated import
